refactoring by moving app files into good location
