+++ infra/src/main/kotlin/fr/sacane/jmanager/server/README.md
