+++ infra/src/main/kotlin/fr/sacane/jmanager/server/adapters/LoginTransactionAdapter.kt
@@ -1,14 +1,13 @@
-package fr.sacane.jmanager.infra.adapters
+package fr.sacane.jmanager.server.adapters
 
 import fr.sacane.jmanager.domain.Hash
 import fr.sacane.jmanager.domain.hexadoc.DatasourceAdapter
 import fr.sacane.jmanager.domain.models.*
 import fr.sacane.jmanager.domain.port.serverside.LoginTransactor
-import fr.sacane.jmanager.infra.entity.Login
-import fr.sacane.jmanager.infra.repositories.LoginRepository
-import fr.sacane.jmanager.infra.repositories.UserRepository
+import fr.sacane.jmanager.server.entity.Login
+import fr.sacane.jmanager.server.repositories.LoginRepository
+import fr.sacane.jmanager.server.repositories.UserRepository
 import org.springframework.stereotype.Service
-import java.security.MessageDigest
 import java.time.LocalDateTime
 import java.util.*
