+++ infra/src/main/kotlin/fr/sacane/jmanager/server/adapters/ServerTransactionAdapter.kt
@@ -1,11 +1,11 @@
-package fr.sacane.jmanager.infra.adapters
+package fr.sacane.jmanager.server.adapters
 
 import fr.sacane.jmanager.domain.hexadoc.DatasourceAdapter
 import fr.sacane.jmanager.domain.models.*
 import fr.sacane.jmanager.domain.port.serverside.TransactionRegister
-import fr.sacane.jmanager.infra.entity.CategoryResource
-import fr.sacane.jmanager.infra.repositories.CategoryRepository
-import fr.sacane.jmanager.infra.repositories.UserRepository
+import fr.sacane.jmanager.server.entity.CategoryResource
+import fr.sacane.jmanager.server.repositories.CategoryRepository
+import fr.sacane.jmanager.server.repositories.UserRepository
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
