+++ infra/src/main/kotlin/fr/sacane/jmanager/server/entity/AccountResource.kt
@@ -1,4 +1,4 @@
-package fr.sacane.jmanager.infra.entity
+package fr.sacane.jmanager.server.entity
 
 import fr.sacane.jmanager.domain.hexadoc.DatasourceEntity
 import org.springframework.lang.Nullable
