+++ infra/src/main/kotlin/fr/sacane/jmanager/server/repositories/AccountRepository.kt
@@ -1,6 +1,6 @@
-package fr.sacane.jmanager.infra.repositories
+package fr.sacane.jmanager.server.repositories
 
-import fr.sacane.jmanager.infra.entity.AccountResource
+import fr.sacane.jmanager.server.entity.AccountResource
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
