+++ infra/src/main/kotlin/fr/sacane/jmanager/server/repositories/CategoryRepository.kt
@@ -1,6 +1,6 @@
-package fr.sacane.jmanager.infra.repositories
+package fr.sacane.jmanager.server.repositories
 
-import fr.sacane.jmanager.infra.entity.CategoryResource
+import fr.sacane.jmanager.server.entity.CategoryResource
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.transaction.annotation.Transactional
