+++ infra/src/test/kotlin/fr/sacane/jmanager/server/InfraUserTest.kt
@@ -1,11 +1,11 @@
-package fr.sacane.jmanager.infra
+package fr.sacane.jmanager.server
 
 import fr.sacane.jmanager.domain.models.Password
-import fr.sacane.jmanager.infra.entity.AccountResource
-import fr.sacane.jmanager.infra.entity.SheetResource
-import fr.sacane.jmanager.infra.entity.UserResource
-import fr.sacane.jmanager.infra.repositories.AccountRepository
-import fr.sacane.jmanager.infra.repositories.UserRepository
+import fr.sacane.jmanager.server.entity.AccountResource
+import fr.sacane.jmanager.server.entity.SheetResource
+import fr.sacane.jmanager.server.entity.UserResource
+import fr.sacane.jmanager.server.repositories.AccountRepository
+import fr.sacane.jmanager.server.repositories.UserRepository
 import org.junit.jupiter.api.Test
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterEach
@@ -13,7 +13,6 @@ import org.junit.jupiter.api.Order
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest
-import org.springframework.test.context.ActiveProfiles
 import org.springframework.test.context.TestPropertySource
